+++ tests/AdagucTests/TestWMS.py
@@ -2141,82 +2141,82 @@ class TestWMS(unittest.TestCase):
             "SERVICE=WMS&VERSION=1.3.0&REQUEST=GetMap&LAYERS=data&WIDTH=360&HEIGHT=180&CRS=EPSG%3A4326&BBOX=-90,-180,90,180&STYLES=auto%2Fnearest&FORMAT=image/png&TRANSPARENT=TRUE&COLORSCALERANGE=0,1&",
             {"ADAGUC_CONFIG": config},
         )
-        self.assertEqual(headers, ["Content-Type:image/png", "Cache-Control:max-age=60"])
+        self.assertEqual(headers, ["Content-Type:image/png", "Content-Encoding:png8a", "Cache-Control:max-age=60"])
 
         status, data, headers = AdagucTestTools().runADAGUCServer(
             "SERVICE=WMS&VERSION=1.3.0&REQUEST=GetMap&LAYERS=data&WIDTH=360&HEIGHT=180&CRS=EPSG%3A4326&BBOX=-90,-180,90,180&STYLES=auto%2Fnearest&FORMAT=image/png&TRANSPARENT=TRUE&COLORSCALERANGE=0,1&time=2017-01-01T00:05:00",
             {"ADAGUC_CONFIG": config},
         )
-        self.assertEqual(headers, ["Content-Type:image/png", "Cache-Control:max-age=60"])
+        self.assertEqual(headers, ["Content-Type:image/png", "Content-Encoding:png8a", "Cache-Control:max-age=60"])
 
         status, data, headers = AdagucTestTools().runADAGUCServer(
             "SERVICE=WMS&VERSION=1.3.0&REQUEST=GetMap&LAYERS=data&WIDTH=360&HEIGHT=180&CRS=EPSG%3A4326&BBOX=-90,-180,90,180&STYLES=auto%2Fnearest&FORMAT=image/png&TRANSPARENT=TRUE&COLORSCALERANGE=0,1&time=2017-01-01T00:05:00&DIM_member=member3",
             {"ADAGUC_CONFIG": config},
         )
-        self.assertEqual(headers, ["Content-Type:image/png", "Cache-Control:max-age=60"])
+        self.assertEqual(headers, ["Content-Type:image/png", "Content-Encoding:png8a", "Cache-Control:max-age=60"])
 
         # If some dims are specfied with current, we expect a shorter caching interval.
         status, data, headers = AdagucTestTools().runADAGUCServer(
             "SERVICE=WMS&VERSION=1.3.0&REQUEST=GetMap&LAYERS=data&WIDTH=360&HEIGHT=180&CRS=EPSG%3A4326&BBOX=-90,-180,90,180&STYLES=auto%2Fnearest&FORMAT=image/png&TRANSPARENT=TRUE&COLORSCALERANGE=0,1&time=current&DIM_member=member3&elevation=5000",
             {"ADAGUC_CONFIG": config},
         )
-        self.assertEqual(headers, ["Content-Type:image/png", "Cache-Control:max-age=60"])
+        self.assertEqual(headers, ["Content-Type:image/png", "Content-Encoding:png8a", "Cache-Control:max-age=60"])
 
         # If some dims are specfied incorrectly, we expect a shorter caching interval.
         status, data, headers = AdagucTestTools().runADAGUCServer(
             "SERVICE=WMS&VERSION=1.3.0&REQUEST=GetMap&LAYERS=data&WIDTH=360&HEIGHT=180&CRS=EPSG%3A4326&BBOX=-90,-180,90,180&STYLES=auto%2Fnearest&FORMAT=image/png&TRANSPARENT=TRUE&COLORSCALERANGE=0,1&time=2017-01-01T00:05:00.000Z&DIM_member=member3&elevation=5000",
             {"ADAGUC_CONFIG": config},
         )
-        self.assertEqual(headers, ["Content-Type:image/png", "Cache-Control:max-age=60"])
+        self.assertEqual(headers, ["Content-Type:image/png", "Content-Encoding:png8a", "Cache-Control:max-age=60"])
 
         # If all dims are specfied, we expect a longer caching interval.
         status, data, headers = AdagucTestTools().runADAGUCServer(
             "SERVICE=WMS&VERSION=1.3.0&REQUEST=GetMap&LAYERS=data&WIDTH=360&HEIGHT=180&CRS=EPSG%3A4326&BBOX=-90,-180,90,180&STYLES=auto%2Fnearest&FORMAT=image/png&TRANSPARENT=TRUE&COLORSCALERANGE=0,1&time=2017-01-01T00:05:00Z&DIM_member=member3&elevation=5000",
             {"ADAGUC_CONFIG": config},
         )
-        self.assertEqual(headers, ["Content-Type:image/png", "Cache-Control:max-age=7200"])
+        self.assertEqual(headers, ["Content-Type:image/png", "Content-Encoding:png8a", "Cache-Control:max-age=7200"])
 
         # If all dims are specfied, but time is without zulue, we expect a shorter caching interval.
         status, data, headers = AdagucTestTools().runADAGUCServer(
             "SERVICE=WMS&VERSION=1.3.0&REQUEST=GetMap&LAYERS=data&WIDTH=360&HEIGHT=180&CRS=EPSG%3A4326&BBOX=-90,-180,90,180&STYLES=auto%2Fnearest&FORMAT=image/png&TRANSPARENT=TRUE&COLORSCALERANGE=0,1&time=2017-01-01T00:05:00&DIM_member=member3&elevation=5000",
             {"ADAGUC_CONFIG": config},
         )
-        self.assertEqual(headers, ["Content-Type:image/png", "Cache-Control:max-age=60"])
+        self.assertEqual(headers, ["Content-Type:image/png", "Content-Encoding:png8a", "Cache-Control:max-age=60"])
 
         # If all dims are specfied, but time contains millieseconds, we expect a shorter caching interval.
         status, data, headers = AdagucTestTools().runADAGUCServer(
             "SERVICE=WMS&VERSION=1.3.0&REQUEST=GetMap&LAYERS=data&WIDTH=360&HEIGHT=180&CRS=EPSG%3A4326&BBOX=-90,-180,90,180&STYLES=auto%2Fnearest&FORMAT=image/png&TRANSPARENT=TRUE&COLORSCALERANGE=0,1&time=2017-01-01T00:05:00.000Z&DIM_member=member3&elevation=5000",
             {"ADAGUC_CONFIG": config},
         )
-        self.assertEqual(headers, ["Content-Type:image/png", "Cache-Control:max-age=60"])
+        self.assertEqual(headers, ["Content-Type:image/png", "Content-Encoding:png8a", "Cache-Control:max-age=60"])
 
         # If all dims are specfied, but time is current, we expect a shorter caching interval.
         status, data, headers = AdagucTestTools().runADAGUCServer(
             "SERVICE=WMS&VERSION=1.3.0&REQUEST=GetMap&LAYERS=data&WIDTH=360&HEIGHT=180&CRS=EPSG%3A4326&BBOX=-90,-180,90,180&STYLES=auto%2Fnearest&FORMAT=image/png&TRANSPARENT=TRUE&COLORSCALERANGE=0,1&time=current&DIM_member=member3&elevation=5000",
             {"ADAGUC_CONFIG": config},
         )
-        self.assertEqual(headers, ["Content-Type:image/png", "Cache-Control:max-age=60"])
+        self.assertEqual(headers, ["Content-Type:image/png", "Content-Encoding:png8a", "Cache-Control:max-age=60"])
 
         # If all dims are specfied except time, we expect a shorter caching interval.
         status, data, headers = AdagucTestTools().runADAGUCServer(
             "SERVICE=WMS&VERSION=1.3.0&REQUEST=GetMap&LAYERS=data&WIDTH=360&HEIGHT=180&CRS=EPSG%3A4326&BBOX=-90,-180,90,180&STYLES=auto%2Fnearest&FORMAT=image/png&TRANSPARENT=TRUE&COLORSCALERANGE=0,1&DIM_member=member3&elevation=5000",
             {"ADAGUC_CONFIG": config},
         )
-        self.assertEqual(headers, ["Content-Type:image/png", "Cache-Control:max-age=60"])
+        self.assertEqual(headers, ["Content-Type:image/png", "Content-Encoding:png8a", "Cache-Control:max-age=60"])
 
         # If all dims are specfied except member, we expect a shorter caching interval.
         status, data, headers = AdagucTestTools().runADAGUCServer(
             "SERVICE=WMS&VERSION=1.3.0&REQUEST=GetMap&LAYERS=data&WIDTH=360&HEIGHT=180&CRS=EPSG%3A4326&BBOX=-90,-180,90,180&STYLES=auto%2Fnearest&FORMAT=image/png&TRANSPARENT=TRUE&COLORSCALERANGE=0,1&time=2017-01-01T00:05:00Z&elevation=5000",
             {"ADAGUC_CONFIG": config},
         )
-        self.assertEqual(headers, ["Content-Type:image/png", "Cache-Control:max-age=60"])
+        self.assertEqual(headers, ["Content-Type:image/png", "Content-Encoding:png8a", "Cache-Control:max-age=60"])
 
         # If all dims are specfied except height, we expect a shorter caching interval.
         status, data, headers = AdagucTestTools().runADAGUCServer(
             "SERVICE=WMS&VERSION=1.3.0&REQUEST=GetMap&LAYERS=data&WIDTH=360&HEIGHT=180&CRS=EPSG%3A4326&BBOX=-90,-180,90,180&STYLES=auto%2Fnearest&FORMAT=image/png&TRANSPARENT=TRUE&COLORSCALERANGE=0,1&time=2017-01-01T00:05:00Z&DIM_member=member3&",
             {"ADAGUC_CONFIG": config},
         )
-        self.assertEqual(headers, ["Content-Type:image/png", "Cache-Control:max-age=60"])
+        self.assertEqual(headers, ["Content-Type:image/png", "Content-Encoding:png8a", "Cache-Control:max-age=60"])
 
         #### From here, dimension 'member' is fixed to 'member4'
         # If we query member with same value as what it is fixed to, we get long cache
@@ -2224,28 +2224,28 @@ class TestWMS(unittest.TestCase):
             "SERVICE=WMS&VERSION=1.3.0&REQUEST=GetMap&LAYERS=data_with_fixed&WIDTH=360&HEIGHT=180&CRS=EPSG%3A4326&BBOX=-90,-180,90,180&STYLES=auto%2Fnearest&FORMAT=image/png&TRANSPARENT=TRUE&COLORSCALERANGE=0,1&time=2017-01-01T00:05:00Z&DIM_member=member4&elevation=5000",
             {"ADAGUC_CONFIG": config},
         )
-        self.assertEqual(headers, ["Content-Type:image/png", "Cache-Control:max-age=7200"])
+        self.assertEqual(headers, ["Content-Type:image/png", "Content-Encoding:png8a", "Cache-Control:max-age=7200"])
 
         # If we query member with different value as what it is fixed to, we get long cache. Fixed values wins from queried value.
         status, data, headers = AdagucTestTools().runADAGUCServer(
             "SERVICE=WMS&VERSION=1.3.0&REQUEST=GetMap&LAYERS=data_with_fixed&WIDTH=360&HEIGHT=180&CRS=EPSG%3A4326&BBOX=-90,-180,90,180&STYLES=auto%2Fnearest&FORMAT=image/png&TRANSPARENT=TRUE&COLORSCALERANGE=0,1&time=2017-01-01T00:05:00Z&DIM_member=member2&elevation=5000",
             {"ADAGUC_CONFIG": config},
         )
-        self.assertEqual(headers, ["Content-Type:image/png", "Cache-Control:max-age=7200"])
+        self.assertEqual(headers, ["Content-Type:image/png", "Content-Encoding:png8a", "Cache-Control:max-age=7200"])
 
         # If we don't provide query at all for member, we should get long cache
         status, data, headers = AdagucTestTools().runADAGUCServer(
             "SERVICE=WMS&VERSION=1.3.0&REQUEST=GetMap&LAYERS=data_with_fixed&WIDTH=360&HEIGHT=180&CRS=EPSG%3A4326&BBOX=-90,-180,90,180&STYLES=auto%2Fnearest&FORMAT=image/png&TRANSPARENT=TRUE&COLORSCALERANGE=0,1&time=2017-01-01T00:05:00Z&elevation=5000",
             {"ADAGUC_CONFIG": config},
         )
-        self.assertEqual(headers, ["Content-Type:image/png", "Cache-Control:max-age=7200"])
+        self.assertEqual(headers, ["Content-Type:image/png", "Content-Encoding:png8a", "Cache-Control:max-age=7200"])
 
         # Unfixed, unprovided dimension should still result in short cache
         status, data, headers = AdagucTestTools().runADAGUCServer(
             "SERVICE=WMS&VERSION=1.3.0&REQUEST=GetMap&LAYERS=data_with_fixed&WIDTH=360&HEIGHT=180&CRS=EPSG%3A4326&BBOX=-90,-180,90,180&STYLES=auto%2Fnearest&FORMAT=image/png&TRANSPARENT=TRUE&COLORSCALERANGE=0,1&time=2017-01-01T00:05:00Z",
             {"ADAGUC_CONFIG": config},
         )
-        self.assertEqual(headers, ["Content-Type:image/png", "Cache-Control:max-age=60"])
+        self.assertEqual(headers, ["Content-Type:image/png", "Content-Encoding:png8a", "Cache-Control:max-age=60"])
 
     def test_WMSGetMap_IrregularGrid_1Dimensional_latlon(self):
         """